--- src/main/java/com/codegym/hotel/booking/repository/MessageRepository.java
@@ -1,4 +1,0 @@
-package com.codegym.hotel.booking.repository;
-
-public interface MessageRepository {
-}
