--- src/main/java/com/codegym/hotel/booking/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package com.codegym.hotel.booking.repository;
-
-public interface UserRepository {
-}
