--- src/main/java/com/codegym/hotel/booking/repository/impl/MessageRepositoryImpl.java
@@ -1,6 +1,0 @@
-package com.codegym.hotel.booking.repository.impl;
-
-import com.codegym.hotel.booking.repository.MessageRepository;
-
-public class MessageRepositoryImpl implements MessageRepository {
-}
