--- src/main/java/com/codegym/hotel/booking/repository/impl/RoleRepositoryImpl.java
@@ -1,6 +1,0 @@
-package com.codegym.hotel.booking.repository.impl;
-
-import com.codegym.hotel.booking.repository.RoleRepository;
-
-public class RoleRepositoryImpl implements RoleRepository {
-}
