--- src/main/java/com/codegym/hotel/booking/repository/impl/UserRepositoryImpl.java
@@ -1,6 +1,0 @@
-package com.codegym.hotel.booking.repository.impl;
-
-import com.codegym.hotel.booking.repository.UserRepository;
-
-public class UserRepositoryImpl implements UserRepository {
-}
