--- src/main/java/com/codegym/hotel/booking/service/MessageService.java
@@ -1,4 +1,0 @@
-package com.codegym.hotel.booking.service;
-
-public interface MessageService {
-}
