--- src/main/java/com/codegym/hotel/booking/service/UserService.java
@@ -1,4 +1,0 @@
-package com.codegym.hotel.booking.service;
-
-public interface UserService {
-}
