--- src/main/java/com/codegym/hotel/booking/service/impl/MessageServiceImpl.java
@@ -1,6 +1,0 @@
-package com.codegym.hotel.booking.service.impl;
-
-import com.codegym.hotel.booking.service.MessageService;
-
-public class MessageServiceImpl implements MessageService {
-}
