--- src/main/java/com/codegym/hotel/booking/service/impl/RoleServiceImpl.java
@@ -1,6 +1,0 @@
-package com.codegym.hotel.booking.service.impl;
-
-import com.codegym.hotel.booking.service.RoleService;
-
-public class RoleServiceImpl implements RoleService {
-}
