--- src/main/java/com/codegym/hotel/booking/service/impl/UserServiceImpl.java
@@ -1,6 +1,0 @@
-package com.codegym.hotel.booking.service.impl;
-
-import com.codegym.hotel.booking.service.UserService;
-
-public class UserServiceImpl implements UserService {
-}
