+++ src/main/java/com/hotel/booking/repository/MessageRepository.java
@@ -1,0 +1,4 @@
+package com.hotel.booking.repository;
+
+public interface MessageRepository {
+}
