+++ src/main/java/com/hotel/booking/repository/RoleRepository.java
@@ -1,0 +1,4 @@
+package com.hotel.booking.repository;
+
+public interface RoleRepository {
+}
