+++ src/main/java/com/hotel/booking/repository/impl/MessageRepositoryImpl.java
@@ -1,0 +1,6 @@
+package com.hotel.booking.repository.impl;
+
+import com.hotel.booking.repository.MessageRepository;
+
+public class MessageRepositoryImpl implements MessageRepository {
+}
