+++ src/main/java/com/hotel/booking/repository/impl/RoleRepositoryImpl.java
@@ -1,0 +1,6 @@
+package com.hotel.booking.repository.impl;
+
+import com.hotel.booking.repository.RoleRepository;
+
+public class RoleRepositoryImpl implements RoleRepository {
+}
