+++ src/main/java/com/hotel/booking/repository/impl/UserRepositoryImpl.java
@@ -1,0 +1,6 @@
+package com.hotel.booking.repository.impl;
+
+import com.hotel.booking.repository.UserRepository;
+
+public class UserRepositoryImpl implements UserRepository {
+}
