+++ src/main/java/com/hotel/booking/service/RoleService.java
@@ -1,0 +1,4 @@
+package com.hotel.booking.service;
+
+public interface RoleService {
+}
