+++ src/main/java/com/hotel/booking/service/impl/MessageServiceImpl.java
@@ -1,0 +1,6 @@
+package com.hotel.booking.service.impl;
+
+import com.hotel.booking.service.MessageService;
+
+public class MessageServiceImpl implements MessageService {
+}
