+++ src/main/java/com/hotel/booking/service/impl/RoleServiceImpl.java
@@ -1,0 +1,6 @@
+package com.hotel.booking.service.impl;
+
+import com.hotel.booking.service.RoleService;
+
+public class RoleServiceImpl implements RoleService {
+}
