+++ src/main/java/com/hotel/booking/service/impl/UserServiceImpl.java
@@ -1,0 +1,6 @@
+package com.hotel.booking.service.impl;
+
+import com.hotel.booking.service.UserService;
+
+public class UserServiceImpl implements UserService {
+}
